refactor: extract validate funcition
